requirements.txt
Django==5.1.4
djangorestframework==3.14
psycopg2==2.9
sqlalchemy==1.4
alembic==1.11
